# Optimization of wfdb.io.annotation.field2bytes function 